Add the version whose relation to the others I can't determine.
These files were in Fractals/RigoNathanColor in the backup.

This version definitely shares a common ancestor with all three of the
linearly-related versions -- starting with the version in
Fractals/FinalWithGroup in the backup -- but based on the diff I can't
tell if the common ancestor is missing, or is that version:

? some printf dubugs disappear in the diff while others get commented
  out. The printfs that disappear are written in a much more informal
  style than the ones that get commented out.  This makes me wonder if
  there was a common ancestor did not have the informal printfs.  On
  the other hand, it could be that we tried to remove all of the
  "unprofessional" comments when creating this version.

? there are arrays 'aHistory' and 'bHistory' which disappear in the
  diff.  However, these arrays seem to be ironic optimization, so
  maybe the were removed to make the code faster: we do a linear
  lookup in them in an attempt to short circuit the iteration if our
  value recurs.  However, this linear check changes the main loop from
  'Theta(n)' to 'Theta(n^2)' complexity :P

In any case, this version is significantly different from the others
-- using a totally different color choosing scheme -- and so I'll want
to compare this version and the newest of the linearly related
versions.