Use libpng++ instead of "nrd" format.
With this commit we're generating images again, now in PNG format.
The use of libpng++ is dead simple here!

I'm going to change 'julia.cpp' to use command line args instead of
prompting, and then do more tests.  So far, it seems that "slow mode"
is actually faster, and is responsible for all the interesting "tails"
in the generated image. I guess this means that most non-diverging
points are recurrent!